server/requirements.txt
Flask
Flask-JWT-Extended
Flask-Migrate
Flask-SQLAlchemy
SQLAlchemy
psycopg2-binary
alembic
python-dotenv
marshmallow
marshmallow-sqlalchemy
Flask-CORS
gunicorn
requests
pytest
pytest-flask



Flask==2.3.3
Flask-CORS==4.0.0
psycopg2-binary==2.9.7
python-dotenv==1.0.0
bcrypt==4.0.1
PyJWT==2.8.0
Werkzeug==2.3.7
gunicorn==21.2.0
pytest==7.4.2